fix: close all leftover sweep items — budget TOCTOU, posix hardening + CI proof, pdf temp-file, modal parity
Native:
- BGC shared budget check-then-decrement TOCTOU -> bgcBudgetTryAcquire
  (fetch_sub commit point, never negative; contention test: budget 1 x 8
  workers advances exactly once)
- disk_reader_posix parity with win: noteBadRead implemented (was no-op),
  XTS key writes/reads under ioMutex_, self-copy guard, setRaidBackend no
  longer resets the backend it just set (closeDriveUnlocked call removed)
- CI native-posix-syntax job (ubuntu, g++ -fsyntax-only over 54
  platform-independent TUs) — already proved itself by catching 3 real
  hidden portability bugs locally (missing virtual_raid.h include,
  missing <cstring>, ull literal template deduction), all fixed
- test temp DB paths now PID-unique (concurrent byteback_tests instances
  collided during multi-lane sweeps)
- utf8SanitizeLenientPreserving: every invalid byte becomes one '?' (length
  1:1, strict-valid output) — snippet offsets can no longer drift; names/
  paths keep the existing sanitizer; overlong/surrogate/truncated tests

TS:
- SsdTrimModal had the ConfirmModal focus-churn pattern -> onCancelRef fix
- ConfirmModal: module-level modal stack — Escape goes to topmost only
- jsdom + @testing-library devDeps; ConfirmModal.test.tsx (5 tests: escape
  once, tab trap incl. shift-wrap, backdrop click, focus restore, stacked
  topmost routing)
- setPermissionCheckHandler companion (same policy as request handler)
- export-report-pdf: data: URL (~2MB cap) -> temp html file + loadFile,
  unlink in finally after win.destroy; new e2e/report-pdf.spec.ts drives
  the real handler and asserts no temp html leak (the only e2e that
  exercised this handler at all)

Gates: native 523/523, vitest 213/213, e2e 15/15, tsc x3 clean

+++ native/src/carver/parsers/bounded_size_parsers.cpp
@@ -73,7 +73,10 @@ StructuralParseResult parseTiff(const uint8_t* data, size_t size) {
     while (ifd != 0 && ifd + 2 <= size && chain < 8) {
         const uint16_t count = rd16(data + ifd, le);
         if (count == 0 || count > 4096 || ifd + 2 + 12ull * count + 4 > size) return r;
-        end = std::max(end, static_cast<uint64_t>(ifd) + 2 + 12ull * count + 4);
+        // CA-056: keep every operand uint64_t — on Linux uint64_t is
+        // `unsigned long`, so mixing in `12ull` (unsigned long long) broke
+        // std::max's template deduction (MSVC has a single 64-bit type).
+        end = std::max(end, static_cast<uint64_t>(ifd) + 2 + 12 * static_cast<uint64_t>(count) + 4);
         const uint8_t* ent = data + ifd + 2;
         for (uint16_t i = 0; i < count; ++i, ent += 12) {
             const uint16_t tag = rd16(ent, le);
